Moved the CyclicBuffer to the Collections dir
